+++ metaheuristics/algorithms/genetic_algorithms/rkga.cpp
@@ -187,11 +187,11 @@ template<class G>
 void rkga<G>::print_performance() const {
 	std::cout << "RKGA algorithm performance:" << std::endl;
 	std::cout << "    Total generation average:          " << GA<G>::total_time << std::endl;
-	std::cout << "    Average generation average:        " << GA<G>::total_time/GA<G>::N_GEN << std::endl;
+	std::cout << "    Average generation average:        " << GA<G>::total_time/static_cast<double>(GA<G>::N_GEN) << std::endl;
 	std::cout << "    Total mutant generation time:      " << GA<G>::mutant_time << std::endl;
-	std::cout << "    Average mutant generation time:    " << GA<G>::mutant_time/GA<G>::N_GEN << std::endl;
+	std::cout << "    Average mutant generation time:    " << GA<G>::mutant_time/static_cast<double>(GA<G>::N_GEN) << std::endl;
 	std::cout << "    Total crossover generation time:   " << GA<G>::crossover_time << std::endl;
-	std::cout << "    Average crossover generation time: " << GA<G>::crossover_time/GA<G>::N_GEN << std::endl;
+	std::cout << "    Average crossover generation time: " << GA<G>::crossover_time/static_cast<double>(GA<G>::N_GEN) << std::endl;
 	std::cout << std::endl;
 }
